Add a RawOverflowPage type and tests
The idea is it's much the same as RawPage, but not expecting to be in
the page binary format, and allowing cheap O(1) non-copying conversion
from serialised values as RawBytes.

It should also help clarify things to have a type distinction between
normal RawPage (in the page format) and overflow pages (opaque data).