Change PageAcc1.singletonPage to return (RawPage, [RawOverflowPage])
This makes the types clearer. It also means the result is now always a
whole number of pages, rather than an arbitrary size that would need
padding out to page size later. Doing it here gives us clearer types
and simplifies the I/O.

Update the tests. Some of the changes are non-trivial due to the new API
doing the padding out to whole page size. Arguably the result is
cleaner, since we now test (roughly) that v = v' <> suffix.