Resolve module naming issue
When it comes to list-like data, this module uses vectors (ordinary and
primitive ones) almost exclusively; so it makes sense to treat them as a
default. For ordinary vectors, we achieve this by not using
qualification at all. For primitive vectors, the most sensible solution
is then to have the qualifier state what distinguishes them from
ordinary vectors: that they are primitive. Hence, we prefer the
qualifier `Primitive` for primitive vectors in the context of this
module.